moving constants
